+++ src/core/state/requests/index.ts
@@ -2,6 +2,7 @@ import create from 'zustand';
 
 import { ProviderRequestPayload } from '../../transports/providerRequestTransport';
 import { createStore } from '../internal/createStore';
+import { withSelectors } from '../internal/withSelectors';
 
 export interface PendingRequestsStore {
   pendingRequests: ProviderRequestPayload[];
@@ -31,4 +32,6 @@ export const pendingRequestStore = createStore<PendingRequestsStore>(
   },
 );
 
-export const usePendingRequestStore = create(pendingRequestStore);
+export const usePendingRequestStore = withSelectors(
+  create(pendingRequestStore),
+);
